requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
Django==3.1.4
django-cors-headers==3.6.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.8
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0